Fix zeroed EndsAt timestamp
This commit fixes a bug where firing alerts had zeroed EndsAt
timestamps in notifications. This bug was introduced in another
PR (#1686) to fix issues where alerts are resolved during a flush.

This commit takes a different approach where instead of removing
the EndsAt timestamp from the alert, it uses the ResolvedAt and
StatusAt methods with the timestamp in the context, not the current
time.

Signed-off-by: George Robinson <george.robinson@grafana.com>